chore: add pinned requirements.txt for test harness
