+++ requirements.txt
@@ -1,0 +1,4 @@
+
+pytest==8.3.5
+psycopg2-binary==2.9.7
+python-dotenv==1.0.0
